This module solves the following problems:
For Windows OS When the uploadFile method is executed, the Japanese file is not converted correctly and an error occurs.
This is an error that occurs because Japanese is recognized as Chinese. Solved this problem.

The changes from the official version are as follows.

Modify the chkChinese method in the OssUtil.php file
Addition of hasSimplifiedChinese method
Addition of hasTraditionalChinese method
Add the following files to the /src/OSS/Core directory

- SimplifiedChinese.txt
- Traditional Chinese.txt

### About the test
Only unit tests of the following files will be run.
The files to be changed are as follows.

- OssClientImageTest.php
- OssClientObjectTest.php